Rewrite the TrackScheme actions with the ui-behaviour framework.
