+++ src/main/java/fiji/plugin/trackmate/visualization/trackscheme/TrackSchemeFrame.java
@@ -41,8 +41,6 @@ import com.mxgraph.swing.handler.mxRubberband;
 
 import fiji.plugin.trackmate.Logger;
 import fiji.plugin.trackmate.gui.displaysettings.DisplaySettings;
-import fiji.plugin.trackmate.util.TrackNavigator;
-import fiji.plugin.trackmate.visualization.TrackMateModelView;
 
 public class TrackSchemeFrame extends JFrame
 {
@@ -114,13 +112,6 @@ public class TrackSchemeFrame extends JFrame
 		final JSplitPane splitPane = new JSplitPane( JSplitPane.HORIZONTAL_SPLIT, inner, graphComponent );
 		splitPane.setDividerLocation( 170 );
 		getContentPane().add( splitPane, BorderLayout.CENTER );
-
-		final TrackSchemeKeyboardHandler keyboardHandler = new TrackSchemeKeyboardHandler( trackScheme.getModel(), graphComponent, new TrackNavigator( trackScheme.getModel(), trackScheme.getSelectionModel() ) );
-		keyboardHandler.installKeyboardActions( graphComponent );
-		keyboardHandler.installKeyboardActions( infoPane );
-
-		// Undo / redo
-		TrackMateModelView.registerUndoShortcut( this, trackScheme.getModel() );
 	}
 
 	/*
